requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0